requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
PyJWT[crypto]==2.9.0
pydantic==2.9.1
pydantic-settings==2.6.0
SQLAlchemy==2.0.35
psycopg[binary]==3.2.3
pgvector==0.3.5
openai==1.51.2
cryptography==42.0.5
tenacity==8.2.3
Pillow==10.4.0
stripe==7.8.0
redis==5.0.1
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
pywebpush==1.14.0

# Development and linting tools
ruff==0.7.4
black==24.10.0
pytest==8.3.4

